@@ -1,4 +1,26 @@
 <script>
+	let {
+		left_01,
+		left_02,
+		left_03,
+		left_04,
+		left_05,
+		left_06,
+		left_07,
+		left_08,
+		left_09,
+		left_10,
+		right_01,
+		right_02,
+		right_03,
+		right_04,
+		right_05,
+		right_06,
+		right_07,
+		right_08,
+		right_09,
+		right_10
+	} = $props();
 </script>
 
 <nav class="nav_margin">
@@ -11,35 +33,35 @@
     is-half-fullhd"
 		>
 			<div class="container_left">
-				{#if $$slots.left_01}
-					<span class="slot_padding"><slot name="left_01" /></span>
+				{#if left_01}
+					<span class="slot_padding">{@render left_01?.()}</span>
 				{/if}
-				{#if $$slots.left_02}
-					<span class="slot_padding"><slot name="left_02" /></span>
+				{#if left_02}
+					<span class="slot_padding">{@render left_02?.()}</span>
 				{/if}
-				{#if $$slots.left_03}
-					<span class="slot_padding"><slot name="left_03" /></span>
+				{#if left_03}
+					<span class="slot_padding">{@render left_03?.()}</span>
 				{/if}
-				{#if $$slots.left_04}
-					<span class="slot_padding"><slot name="left_04" /></span>
+				{#if left_04}
+					<span class="slot_padding">{@render left_04?.()}</span>
 				{/if}
-				{#if $$slots.left_05}
-					<span class="slot_padding"><slot name="left_05" /></span>
+				{#if left_05}
+					<span class="slot_padding">{@render left_05?.()}</span>
 				{/if}
-				{#if $$slots.left_06}
-					<span class="slot_padding"><slot name="left_06" /></span>
+				{#if left_06}
+					<span class="slot_padding">{@render left_06?.()}</span>
 				{/if}
-				{#if $$slots.left_07}
-					<span class="slot_padding"><slot name="left_07" /></span>
+				{#if left_07}
+					<span class="slot_padding">{@render left_07?.()}</span>
 				{/if}
-				{#if $$slots.left_08}
-					<span class="slot_padding"><slot name="left_08" /></span>
+				{#if left_08}
+					<span class="slot_padding">{@render left_08?.()}</span>
 				{/if}
-				{#if $$slots.left_09}
-					<span class="slot_padding"><slot name="left_09" /></span>
+				{#if left_09}
+					<span class="slot_padding">{@render left_09?.()}</span>
 				{/if}
-				{#if $$slots.left_10}
-					<span class="slot_padding"><slot name="left_10" /></span>
+				{#if left_10}
+					<span class="slot_padding">{@render left_10?.()}</span>
 				{/if}
 			</div>
 		</div>
@@ -52,35 +74,35 @@
     is-half-fullhd"
 		>
 			<div class="container_right">
-				{#if $$slots.right_10}
-					<span class="slot_padding"><slot name="right_10" /></span>
+				{#if right_10}
+					<span class="slot_padding">{@render right_10?.()}</span>
 				{/if}
-				{#if $$slots.right_09}
-					<span class="slot_padding"><slot name="right_09" /></span>
+				{#if right_09}
+					<span class="slot_padding">{@render right_09?.()}</span>
 				{/if}
-				{#if $$slots.right_08}
-					<span class="slot_padding"><slot name="right_08" /></span>
+				{#if right_08}
+					<span class="slot_padding">{@render right_08?.()}</span>
 				{/if}
-				{#if $$slots.right_07}
-					<span class="slot_padding"><slot name="right_07" /></span>
+				{#if right_07}
+					<span class="slot_padding">{@render right_07?.()}</span>
 				{/if}
-				{#if $$slots.right_06}
-					<span class="slot_padding"><slot name="right_06" /></span>
+				{#if right_06}
+					<span class="slot_padding">{@render right_06?.()}</span>
 				{/if}
-				{#if $$slots.right_05}
-					<span class="slot_padding"><slot name="right_05" /></span>
+				{#if right_05}
+					<span class="slot_padding">{@render right_05?.()}</span>
 				{/if}
-				{#if $$slots.right_04}
-					<span class="slot_padding"><slot name="right_04" /></span>
+				{#if right_04}
+					<span class="slot_padding">{@render right_04?.()}</span>
 				{/if}
-				{#if $$slots.right_03}
-					<span class="slot_padding"><slot name="right_03" /></span>
+				{#if right_03}
+					<span class="slot_padding">{@render right_03?.()}</span>
 				{/if}
-				{#if $$slots.right_02}
-					<span class="slot_padding"><slot name="right_02" /></span>
+				{#if right_02}
+					<span class="slot_padding">{@render right_02?.()}</span>
 				{/if}
-				{#if $$slots.right_01}
-					<span class="slot_padding"><slot name="right_01" /></span>
+				{#if right_01}
+					<span class="slot_padding">{@render right_01?.()}</span>
 				{/if}
 			</div>
 		</div>
