@@ -1,6 +1,23 @@
 <script>
-import { createEventDispatcher } from 'svelte';
-	export let classSize = 'is-small';
+	//import { createEventDispatcher } from 'svelte';
+	let {
+		classSize = $bindable('is-small'),
+		tabs = $bindable([
+			{
+				label: 'Pictures',
+				classIcon: 'fas fa-picture',
+				slot: 'picture',
+				isActive: false,
+				disabled: true
+			},
+			{ label: 'Calcs', classIcon: 'fas fa-picture', isActive: true },
+			{ label: 'Film', classIcon: 'fas fa-film', isActive: false, disabled: true }
+		]),
+		active = $bindable(0),
+		children
+	} = $props();
+	//export let classSize = 'is-small';
+	/*
 	export let tabs = [
 		{
 			label: 'Pictures',
@@ -12,14 +29,20 @@ import { createEventDispatcher } from 'svelte';
 		{ label: 'Calcs', classIcon: 'fas fa-picture',  isActive: true },
 		{ label: 'Film', classIcon: 'fas fa-film',  isActive: false, disabled: true }
 	];
+	*/
 	// slot no used
-	export let active = 0;
+	//	export let active = 0;
 
-	const dispatch = createEventDispatcher();
+	//	const dispatch = createEventDispatcher();
 
 	const withIcons = (currentValue) => currentValue.classIcon && currentValue.classIcon.length > 3;
 
-	$: active, changeTab();
+	//$: active, changeTab();
+	$inspect(active).with((type) => {
+		if (type === 'update' || type === 'init') {
+			changeTab();
+		}
+	});
 
 	function changeTab() {
 		//console.log('changeTab');
@@ -45,7 +68,7 @@ import { createEventDispatcher } from 'svelte';
 			});
 
 			//console.log(exists_active, last_active);
-			
+
 			if (!exists_active && tabs[last_active]) {
 				tabs[last_active].isActive = true;
 			}
@@ -57,11 +80,11 @@ import { createEventDispatcher } from 'svelte';
 	<ul>
 		{#each tabs as item, i}
 			<li class={item.isActive ? 'is-active' : ''}>
-				<!-- svelte-ignore a11y-missing-attribute -->
-				<!-- svelte-ignore a11y-click-events-have-key-events -->
-				<!-- svelte-ignore a11y-no-static-element-interactions -->
+				<!-- svelte-ignore a11y_click_events_have_key_events -->
+				<!-- svelte-ignore a11y_no_static_element_interactions -->
+				<!-- svelte-ignore a11y_missing_attribute -->
 				<a
-					on:click={() => {
+					onclick={() => {
 						/*
 						tabs = tabs.map((t) => {
 							t.isActive = false;
@@ -77,7 +100,8 @@ import { createEventDispatcher } from 'svelte';
 						if (!item.disabled) {
 							// Solo dispara cuando no está deshabilitado
 							console.log({ label: item.label, index: i });
-							dispatch('select',{ label: item.label, index: i });
+							//dispatch('select',{ label: item.label, index: i });
+							onselect({ label: item.label, index: i });
 						}
 					}}
 				>
@@ -85,7 +109,7 @@ import { createEventDispatcher } from 'svelte';
 						<span class="icon"><i class="fa-solid fa-ban" aria-hidden="true"></i></span>
 						<span>{item.label}</span>
 					{:else if tabs.every(withIcons)}
-						<span class="icon"><i class={item.classIcon} aria-hidden="true" /></span>
+						<span class="icon"><i class={item.classIcon} aria-hidden="true"></i></span>
 						<span>{item.label}</span>
 					{:else}
 						{item.label}
@@ -95,4 +119,4 @@ import { createEventDispatcher } from 'svelte';
 		{/each}
 	</ul>
 </div>
-<slot />
+{@render children?.()}
