@@ -1,52 +1,42 @@
 <script>
-	export let value;
-	export let props;
-	export let row = {}; // Required by no user directly
-	let DC001 = false;
+	let {
+		value = $bindable(),
+		onclick_cell = () => {},
+		row = $bindable(),
+		format = 'yyyy-MM-dd HH:mm:ss',
+		fromFormat = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'",
+		HighlightIsntToday = false,
+		editInline = false,
+		css_cell = ''
+	} = $props();
+
+	let DC001 = $state(false);
 	import { DateTime } from 'luxon';
-	let format = 'yyyy-MM-dd HH:mm:ss';
-	let HighlightIsntToday = false;
-	let fromFormat = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'";
-	let value_formated = '';
-
-	if (props) {
-		if (props.format) {
-			format = props.format;
-		}
-
-		if (props.fromFormat) {
-			fromFormat = props.fromFormat;
-		}
-
-		if (props.highlight_isnt_today) {
-			HighlightIsntToday = true;
-		}
-	}
+	let value_formated = $state('');
 
 	function fn_DC01() {
 		try {
 			DC001 =
 				HighlightIsntToday &&
 				DateTime.fromISO(value).toFormat('yyyy-MM-dd') !== DateTime.local().toFormat('yyyy-MM-dd');
-
-			//value_formated = DateTime.fromFormat(value, fromFormat);
 			value_formated = DateTime.fromFormat(value, fromFormat).toLocal().toFormat(format);
 		} catch (error) {
+			//console.error(error);
 			value_formated = value;
 		}
-
-		/*
-      HighlightIsntToday &&
-      !(
-        new Date(value).getFullYear() === new Date().getFullYear() &&
-        new Date(value).getDate() === new Date().getDate()
-      );
-      */
-		//		console.log(fromFormat);
 	}
-	$: value, fn_DC01();
+
+	$inspect(value).with((type) => {
+		//console.log('>>>>>>>>>>>>> ', type);
+		if (type === 'update' || type === 'init') {
+			fn_DC01();
+		}
+		if (editInline) {
+			console.warn('editInline on DateTime isnt enabled.');
+		}
+	});
 </script>
 
-<td on:click class:has-text-danger={DC001}>
+<td onclick={onclick_cell} class:has-text-danger={DC001}>
 	{value_formated}
 </td>
