@@ -1,8 +1,8 @@
 import Boolean from "./Boolean.svelte";
 import TextLimit from "./TextLimit.svelte";
 import DateTime from "./DateTime.svelte";
-import Json from "./JSON.svelte";
+import Json from "./TreeView .svelte";
 import Auto from "./Auto.svelte";
-import BooleanIcon from "./BooleanIcon.svelte";
+//import BooleanIcon from "./BooleanIcon.svelte";
 
-export { Boolean, TextLimit, DateTime, Json, Auto, BooleanIcon };
+export { Boolean, TextLimit, DateTime, Json, Auto };
